more cleanup, plus mix/grab
